[id70] Unit tests coverage

+++ store/models/order_dao.py
@@ -210,7 +210,7 @@ class OrderProduct(Base):
     product_id = Column(Integer, ForeignKey('products.id'), primary_key=True)
     product = relationship('Product', backref=backref('order_product', lazy='dynamic'))
     quantity = Column(Integer)
-    price = Column(Integer, nullable=True)
+    price = Column(DECIMAL, nullable=True)
     #price_id = Column(Integer,ForeignKey('products.price'), nullable=True)
     #price = relationship('Product', backref=backref('order_product', lazy='dynamic'))
